project-folder/requirements.txt
matplotlib
pillow
streamlit
